Initial commit of new basic template.
This is using graphql-core-next, with the addition of schema tree shaking, caching, debugging and various other improvements.

Please see README.ME for more details. I've attempted to make the documentation a partial tutorial, based on my training experiences.

However, it is not complete. The Kubernetes section is taken directly from Almir's original README and I've not tested it myself.